src/pyproject.toml
[project]
name = "sgr-deep-research"
version = "0.1.0"
description = "SGR Deep Research"
authors = [
    {name = "virrius"}
]
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Core dependencies - основные зависимости для работы системы
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    # Search and research - поиск и исследования
    "tavily-python>=0.3.0",
    # Web scraping and content extraction - веб-скрапинг и извлечение контента
    "trafilatura>=1.6.0",
    "youtube-transcript-api>=0.6.0",
    # Configuration and utilities - конфигурация и утилиты
    "PyYAML>=6.0",
    "envyaml>=1.10.0",
    "python-dateutil>=2.8.0",
    "pydantic-settings>=2.10.1",
]



